# Conflicts:
#	CMakeLists.txt
#	src/Engine/EngineInclude.hpp
#	src/Rendering/ImguiRenderer.hpp
#	src/Rendering/Renderers/ClusterRenderer.hpp
#	src/Rendering/RenderingInclude.hpp
#	src/Shaders/glsl/ClusterRendering/light.frag
#	src/Shaders/spirv/ClusterRendering/light.frag.spv

+++ src/WindowAPI/WindowProvider.hpp
@@ -50,7 +50,7 @@ public:
     {
         glfwPollEvents();
     }
-    float GetTime()
+    double GetTime()
     {
         return glfwGetTime();
     }
